@@ -1,7 +1,9 @@
 import { combineReducers} from "redux";
+import input from "./input.js";
 
 const createRoorReducer = () => 
     combineReducers({
+        input
     })
-    
-    export default createRoorReducer
+
+export default createRoorReducer
