@@ -2,11 +2,13 @@ import express from 'express'
 import cors from 'cors'
 import cookieParser from 'cookie-parser'
 import { resolve } from 'path'
+import countryCodeRouter from './routes/countyCode.router.js'
 
 import { Html } from '../client/html.js'
+import { Server } from 'http'
 
 const server = express()
-const PORT = process.env.PORT || 8080
+const PORT = process.env.PORT || 8090
 const __dirname = process.cwd()
 
 const middleware = [
@@ -18,9 +20,7 @@ const middleware = [
 
 middleware.forEach((it) => server.use(it))
 
-server.get('/', (req, res) => {
-    res.send('Express Server')
-})
+server.use('/api', countryCodeRouter)
 
 server.get('/*', (req, res) => {
     const initialState = {
